feat: TypeScript migration - image-validator utility
Migrated image-validator.js to TypeScript:
- URL validation with caching
- Wikimedia URL builder
- Artwork filtering utilities
- Cache warmup for startup

All 220 tests passing.
