+++ server/src/utils/index.ts
@@ -6,3 +6,4 @@ export * from './time';
 export * from './state';
 export * from './validation';
 export * from './data-store';
+export * from './image-validator';
